+++ app/CHANGELOG.md
@@ -1,8 +1,12 @@
+# 3.8.8
+## Feature
+  - เพิ่มการกำหนด charset ใน db.ts (เฉพาะ mysql)
+
 # 3.8.7
 ## Feature
   - เพิ่ม path file config ในคำสั่ง dotenv
   - Update package สำหรับเชื่อมฐานข้อมูล pgSQL และ oracleDB
-  - Update package อื่นๆ เพื่อปิดช่องโหว่ของระบบ
+  - Update package อื่นๆ เพื่อปิดช่องโหว่ของระบบที่อาจจะเกิดขึ้นได้
 
 # 3.8.6
 ## Feature